% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{Get the consensus sequence from a multiple sequence alignment}
\usage{
get_consensus(sequences)
}
\arguments{
\item{sequences}{A set of sequences forming a multiple sequence alignment}
}
\value{
A single sequence, representing the consensus sequence
}
\description{
For each position, the most frequently observed amino acid residue forms the consensus sequence
}
\examples{
get_signisite_zscores(ALIGNMENT$sequence, get_values(ALIGNMENT$fasta_header))

}
