% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signisite_zscores.R
\name{get_signisite_zscores}
\alias{get_signisite_zscores}
\title{Compute the SigniSite z-scores}
\usage{
get_signisite_zscores(sequences, values)
}
\arguments{
\item{sequences}{A set of sequences forming a multiple sequence alignment}

\item{values}{The for each sequence associated numerical value}
}
\value{
A matrix of z-scores, where each score quantifies the strength of the association between a given amino acid residue at a given position and the phenotype
}
\description{
Compute the SigniSite z-scores
}
\examples{
get_signisite_zscores(ALIGNMENT$sequence, get_values(ALIGNMENT$fasta_header))

}
