% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Extract white space separated end-placed values from a set of FASTA headers}
\usage{
get_values(fasta_header)
}
\arguments{
\item{fasta_header}{A vector of FASTA headers formatted like so: ">id info VALUE"}
}
\value{
A numerical vector of the extracted values
}
\description{
Extract white space separated end-placed values from a set of FASTA headers
}
\examples{
get_values(fasta_header = ALIGNMENT$fasta_header)

}
