% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signisite_logo.R
\name{plot_signisite_logo}
\alias{plot_signisite_logo}
\title{Plot the results of a SigniSite analysis as a sequence logo}
\usage{
plot_signisite_logo(z)
}
\arguments{
\item{z}{A matrix of z-scores}
}
\value{
A ggplot object
}
\description{
Plot the results of a SigniSite analysis as a sequence logo
}
\examples{
z_mat = get_signisite_zscores(ALIGNMENT$sequence, get_values(ALIGNMENT$fasta_header))
z_mat_sp = rm_ns_positions(z_mat, alpha = 0.05)
plot_signisite_logo(z_mat_sp)

}
