% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a fasta file}
\usage{
read_fasta(file)
}
\arguments{
\item{file}{A multiple sequence alignment in FASTA format}
}
\value{
A data frame of FASTA headers and sequences
}
\description{
Read a fasta file
}
\examples{
read_fasta(file = 'path/to/my/msa_file.fsa')

}
