% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigniSite-package.R
\docType{package}
\name{SigniSite-package}
\alias{SigniSite-package}
\alias{SigniSite}
\title{Identification of residue-level genotype-phenotype correlations in protein multiple sequence alignments}
\description{
SigniSite performs residue level genotype phenotype correlation in 
protein multiple sequence alignments by identifying amino acid residues 
significantly associated with the phenotype of the data set. 
Input is a protein multiple sequence alignment in FASTA format. The 
phenotype is represented by a real-valued numerical value placed last in 
the identifier of each sequence (white-space separated).
}
\examples{
library("SigniSite")

# Create the matrix of SigniSite z-scores
z = signisite_zmat(file = system.file(package = "SigniSite", "/signisite_alignment.fsa"), 
                   method = 'bonferroni', alpha = 0.05)

# Visualise z-score matrix as a logo plot
plot_signisite_logo(z)
}
\references{
Leon Eyrich Jessen, Ilka Hoof, Ole Lund and Morten Nielsen
}
\author{
Maintainer: Leon Eyrich Jessen <ljess@dtu.dk>
}
\keyword{package}
