% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signisite_zmat.R
\name{signisite_zmat}
\alias{signisite_zmat}
\title{Run the SigniSite workflow}
\usage{
signisite_zmat(file, method = "bonferroni", alpha = 0.05)
}
\arguments{
\item{file}{A multiple sequence alignment in FASTA format}

\item{method}{method for correcting for multiple testing, one of c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"), see ?p.adjust for more. Default: bonferroni}

\item{alpha}{the level of significance. Default: 0.05}
}
\value{
A matrix of SigniSite z-scores, corrected for multiple testing and truncated to significant positions
}
\description{
Run the SigniSite workflow
}
\examples{
signisite_zmat(file = 'path/to/my/msa_file.fsa')

}
