% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_c_factor.R
\name{tie_c_factor}
\alias{tie_c_factor}
\title{Compute the tie correction factor}
\usage{
tie_c_factor(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The computed tie correction factor as a numeric vector of length 1
}
\description{
Compute the tie correction factor
}
\examples{
tie_c_factor(x = round(rnorm(10,0)))

}
