% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_z_matrix.R
\name{correct_z_matrix}
\alias{correct_z_matrix}
\title{Correct a matrix of z-scores for multiple testing}
\usage{
correct_z_matrix(z, method = "bonferroni")
}
\arguments{
\item{z}{A matrix of z-scores}

\item{method}{method for correcting for multiple testing, one of c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"), see ?p.adjust for more. Default: bonferroni}
}
\value{
A matrix of corrected z-scores
}
\description{
Correct a matrix of z-scores for multiple testing
}
\examples{
z_mat = get_signisite_zscores(ALIGNMENT$sequence, get_values(ALIGNMENT$fasta_header))
correct_z_matrix(z = z_mat, method = 'bonferroni')

}
