% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_ns_positions.R
\name{rm_ns_positions}
\alias{rm_ns_positions}
\title{Remove positions where no significant association was identified}
\usage{
rm_ns_positions(z, alpha = 0.05)
}
\arguments{
\item{z}{A matrix of z-scores}
}
\value{
A matrix of z-scores, with non-significant positions removed
}
\description{
Remove positions where no significant association was identified
}
\examples{
z_mat = get_signisite_zscores(ALIGNMENT$sequence, get_values(ALIGNMENT$fasta_header))
rm_ns_positions(z_mat, alpha = 0.05)

}
